+++ demo/src/main/java/uniandes/edu/co/demo/modelo/ProductoPerecedero.java
@@ -5,7 +5,7 @@ import java.util.Date;
 public class ProductoPerecedero extends Producto{
     private Date FechaVencimiento;
 
-    public ProductoPerecedero(Long CodigoBarras,String Nombre, Double PrecioUnitarioVenta,String Presentacion,Integer CantidadPresentacion,String UnidadMedida,Date FechaExpiracion,String tipo,Categoria Categoria,EspecificacionEmpacado EspecificacionEmpacado, Date FechaVencimiento){
+    public ProductoPerecedero(int CodigoBarras,String Nombre, Double PrecioUnitarioVenta,String Presentacion,Integer CantidadPresentacion,String UnidadMedida,Date FechaExpiracion,String tipo,Categoria Categoria,EspecificacionEmpacado EspecificacionEmpacado, Date FechaVencimiento){
         super( CodigoBarras,Nombre,PrecioUnitarioVenta,Presentacion,CantidadPresentacion,UnidadMedida,FechaExpiracion,tipo,Categoria,EspecificacionEmpacado);
         this.FechaVencimiento=FechaVencimiento;
     }
